feat: GameLoop client

+++ client/include/engine/entities/worms/Worm.h
@@ -3,6 +3,7 @@
 #include <SDL_render.h>
 #include "engine/graphics/Animation.h"
 #include "engine/entities/TexturePaths.h"
+#include "engine/utils/WorldScale.h"
 #include <filesystem>
 
 class Worm {
@@ -14,7 +15,7 @@ private:
     BlendMode m_BlendMode;
 
 public:
-    Worm(SDL_Renderer* renderer, int posX, int posY);
+    Worm(SDL_Renderer* renderer, float posX, float posY);
 
     void init();
     void release();
